Rewrite with clone as overlay instead of original as overlay
Because of CSS selector we cannot move the original select
in the DOM (end of body) and still guarantee that the visual aspect
will not change. So we must inverse the concept and use the clone,
with manually cloned computed styles, as an overlay. That makes
events more complex to manage though...